pyproject.toml
[project]
name = "scripts"
version = "0.1.0"
description = "Work summary and automation scripts"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "browser-use>=0.11.1",
    "playwright>=1.57.0",
    # Work Summary dependencies
    "click>=8.1.0",
    "rich>=13.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "aiohttp>=3.9.0",
    "PyGithub>=2.0.0",
    "python-gitlab>=4.0.0",
    "jira>=3.5.0",
    "pyyaml>=6.0.0",
    "python-dateutil>=2.8.0",
    "requests>=2.28.0",
    "python-dotenv>=1.0.0",
    "lightapi>=0.1.16",
    "psycopg2-binary>=2.9.11",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

[project.scripts]
work-summary = "work_summary.cli:main"
